// lib/trpc/schemas/organizations.ts
export const orgCompanyInfoFormSchema = z.object({
  companyDescription: z.string().max(500, 'Description too long').optional(),
  productDescription: z.string().max(500, 'Description too long').optional(),
  serviceDescription: z.string().max(500, 'Description too long').optional(),
});